moved reload services logic to c conf.cfg_instance
